Basic browserification
