+++ front/scripts/Polyfills.js
@@ -1,0 +1,46 @@
+/**
+ * Set Polyfills as a static 'class'.
+ * It's just a static Object, like Math.
+ * It's a pattern which don't need to be extract from main, if it's usage is only to fix
+ * Object.assign existance once
+ */
+export default class Polyfills {
+  /**
+   * Your olf method, wrap in a static method
+   * @example Polyfills.objectAssign()
+   */
+  static objectAssign() {
+    if (!Object.assign) {
+      Object.defineProperty(Object, 'assign', {
+        enumerable: false,
+        configurable: true,
+        writable: true,
+        value: function(target) {
+          'use strict';
+          if (target === undefined || target === null) {
+            throw new TypeError('Cannot convert first argument to object');
+          }
+
+          var to = Object(target);
+          for (var i = 1; i < arguments.length; i++) {
+            var nextSource = arguments[i];
+            if (nextSource === undefined || nextSource === null) {
+              continue;
+            }
+            nextSource = Object(nextSource);
+
+            var keysArray = Object.keys(Object(nextSource));
+            for (var nextIndex = 0, len = keysArray.length; nextIndex < len; nextIndex++) {
+              var nextKey = keysArray[nextIndex];
+              var desc = Object.getOwnPropertyDescriptor(nextSource, nextKey);
+              if (desc !== undefined && desc.enumerable) {
+                to[nextKey] = nextSource[nextKey];
+              }
+            }
+          }
+          return to;
+        }
+      });
+    }
+  }
+}
