+++ front/scripts/TYPE_TO_CSS_CLASS.js
@@ -1,0 +1,22 @@
+const TYPE_TO_CSS_CLASS = [
+  'normal',   // 1
+  'fighting', // 2
+  'flying',   // 3
+  'poison',   // 4
+  'ground',   // 5
+  'rock',     // 6
+  'bug',      // 7
+  'ghost',    // 8
+  'steel',    // 9
+  'fire',     // 10
+  'water',    // 11
+  'grass',    // 12
+  'electric', // 13
+  'psychic',  // 14
+  'ice',      // 15
+  'dragon',   // 16
+  'dark',     // 17
+  'fairy',    // 18
+  'unknown',  // 19
+  'shadow'    // 20
+];
