+++ package.json
@@ -6,10 +6,14 @@
   "devDependencies": {
     "babel-core": "^6.14.0",
     "babel-preset-es2015": "6.13.2",
+    "babelify": "^7.3.0",
+    "browserify": "^13.1.0",
     "gulp": "3.9.1",
     "gulp-babel": "6.1.2",
     "gulp-cssmin": "0.1.7",
-    "gulp-sass": "2.3.2"
+    "gulp-sass": "2.3.2",
+    "vinyl-buffer": "^1.0.0",
+    "vinyl-source-stream": "^1.1.0"
   },
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
